Log lag metric to stdout for Librato
