+++ defaults.env
@@ -79,6 +79,7 @@ EMAIL_QUEUE_ALLOW_UP_TO=3
 
 UPDATE_CTA_EVERY=300
 CHECK_DB_EVERY=600
+CHECK_NPM_SYNC_EVERY=60
 OPTIMIZELY_ID=
 INCLUDE_PIWIK=no
 SENTRY_DSN=
