+++ gratipay/application.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import psycopg2.extras
 
-from . import email, utils
+from . import email, sync_npm, utils
 from .cron import Cron
 from .models import GratipayDB
 from .payday_runner import PaydayRunner
@@ -53,6 +53,8 @@ class Application(object):
         cron(env.update_cta_every, lambda: utils.update_cta(website))
         cron(env.check_db_every, db.self_check, True)
         cron(env.email_queue_flush_every, self.email_queue.flush, True)
+        if website.log_metrics:
+            cron(env.check_npm_sync_every, lambda: sync_npm.check(db), True)
 
 
     def add_event(self, c, type, payload):
