+++ tests/py/fixtures/ConsumeChangeStreamTests.yml
@@ -1,0 +1,22 @@
+interactions:
+- request:
+    body: null
+    headers: {}
+    method: GET
+    uri: https://skimdb.npmjs.com:443/registry
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAEA22RzW7DIBCE732MPftg/sEvgyisXVQbp4CjtlHevTh1JKvqdZaZnf24QXi1yS0I
+        A2ScYqn5CzoIq7d+3VKFgXPdE/YrBZyfstKcd7BdgqtoC37AQIwSXOgOLlueDq3vwK/Lxflq85ZS
+        TBMMo5sLtrxY3m2J320zZUwbRbRUHaz1DTMMN2jB7pgT1WsmCKPq3nwnvbmE0Ua0entS2X1jnP9E
+        tu3x2jRyeo6fFXNyc1PP4TGV6pJv7avL1db4AEO4NlwqSgQxkjKxA9rbj2teXLVXzCWuCQb5uHaJ
+        tWKw/6I5YLTxExnttWgctxjaD1BJQmBcmjF4JGNPzag5E1IH9MaJAPeXH2Iq25KxAQAA
+    headers:
+      cache-control: [must-revalidate]
+      content-encoding: [gzip]
+      content-type: [application/json]
+      strict-transport-security: [max-age=2592000000; includeSubDomains; preload;]
+      transfer-encoding: [chunked]
+    status: {code: 200, message: OK}
+version: 1
